# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Comanda Eletrônica API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type TableStatus = (typeof TableStatus)[keyof typeof TableStatus];

export const TableStatus = {
  available: "available",
  occupied: "occupied",
  reserved: "reserved",
  cleaning: "cleaning",
} as const;

export interface Table {
  id: number;
  number: number;
  capacity: number;
  status: TableStatus;
  guestName?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface CreateTableRequest {
  number: number;
  capacity: number;
}

export interface UpdateTableRequest {
  status?: TableStatus;
  guestName?: string | null;
  capacity?: number;
}

export type OrderStatus = (typeof OrderStatus)[keyof typeof OrderStatus];

export const OrderStatus = {
  pending: "pending",
  preparing: "preparing",
  ready: "ready",
  delivered: "delivered",
  cancelled: "cancelled",
} as const;

export interface OrderItem {
  id: number;
  orderId: number;
  menuItemId: number;
  menuItemName: string;
  menuItemPrice: number;
  quantity: number;
  notes?: string | null;
  subtotal: number;
}

export interface Order {
  id: number;
  tableId: number;
  tableNumber: number;
  status: OrderStatus;
  items: OrderItem[];
  total: number;
  notes?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface CreateOrderItemRequest {
  menuItemId: number;
  quantity: number;
  notes?: string | null;
}

export interface CreateOrderRequest {
  tableId: number;
  items: CreateOrderItemRequest[];
  notes?: string | null;
}

export interface UpdateOrderRequest {
  status?: OrderStatus;
  notes?: string | null;
}

export interface MenuItem {
  id: number;
  name: string;
  description?: string | null;
  price: number;
  categoryId: number;
  categoryName: string;
  available: boolean;
  /** Preparation time in minutes */
  preparationTime?: number | null;
  createdAt: string;
}

export interface CreateMenuItemRequest {
  name: string;
  description?: string | null;
  price: number;
  categoryId: number;
  preparationTime?: number | null;
}

export interface UpdateMenuItemRequest {
  name?: string;
  description?: string | null;
  price?: number;
  categoryId?: number;
  available?: boolean;
  preparationTime?: number | null;
}

export interface Category {
  id: number;
  name: string;
  icon?: string | null;
  sortOrder: number;
}

export interface CreateCategoryRequest {
  name: string;
  icon?: string | null;
  sortOrder: number;
}

export interface TopItem {
  menuItemId: number;
  name: string;
  count: number;
  revenue: number;
}

export interface Stats {
  totalOrders: number;
  totalRevenue: number;
  averageTicket: number;
  activeOrders: number;
  occupiedTables: number;
  totalTables: number;
  topItems: TopItem[];
}

export type ListOrdersParams = {
  status?: ListOrdersStatus;
  tableId?: number;
};

export type ListOrdersStatus =
  (typeof ListOrdersStatus)[keyof typeof ListOrdersStatus];

export const ListOrdersStatus = {
  pending: "pending",
  preparing: "preparing",
  ready: "ready",
  delivered: "delivered",
  cancelled: "cancelled",
} as const;

export type ListMenuItemsParams = {
  categoryId?: number;
  available?: boolean;
};