# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Comanda Eletrônica API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: tables
    description: Mesa operations
  - name: orders
    description: Pedido operations
  - name: menu
    description: Menu item operations
  - name: categories
    description: Category operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /tables:
    get:
      operationId: listTables
      tags: [tables]
      summary: List all tables
      responses:
        "200":
          description: List of tables
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Table"
    post:
      operationId: createTable
      tags: [tables]
      summary: Create a table
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTableRequest"
      responses:
        "201":
          description: Created table
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Table"

  /tables/{id}:
    get:
      operationId: getTable
      tags: [tables]
      summary: Get a table by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Table
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Table"
        "404":
          description: Not found
    patch:
      operationId: updateTable
      tags: [tables]
      summary: Update table status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTableRequest"
      responses:
        "200":
          description: Updated table
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Table"
    delete:
      operationId: deleteTable
      tags: [tables]
      summary: Delete a table
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /tables/{id}/orders:
    get:
      operationId: getTableOrders
      tags: [tables]
      summary: Get all orders for a table
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Orders for the table
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"

  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List all orders
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [pending, preparing, ready, delivered, cancelled]
        - name: tableId
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: List of orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
    post:
      operationId: createOrder
      tags: [orders]
      summary: Create an order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderRequest"
      responses:
        "201":
          description: Created order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"

  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get an order by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
    patch:
      operationId: updateOrder
      tags: [orders]
      summary: Update order status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOrderRequest"
      responses:
        "200":
          description: Updated order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
    delete:
      operationId: deleteOrder
      tags: [orders]
      summary: Cancel/delete an order
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /menu:
    get:
      operationId: listMenuItems
      tags: [menu]
      summary: List all menu items
      parameters:
        - name: categoryId
          in: query
          required: false
          schema:
            type: integer
        - name: available
          in: query
          required: false
          schema:
            type: boolean
      responses:
        "200":
          description: List of menu items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MenuItem"
    post:
      operationId: createMenuItem
      tags: [menu]
      summary: Create a menu item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMenuItemRequest"
      responses:
        "201":
          description: Created menu item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MenuItem"

  /menu/{id}:
    patch:
      operationId: updateMenuItem
      tags: [menu]
      summary: Update a menu item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateMenuItemRequest"
      responses:
        "200":
          description: Updated menu item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MenuItem"
    delete:
      operationId: deleteMenuItem
      tags: [menu]
      summary: Delete a menu item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /categories:
    get:
      operationId: listCategories
      tags: [categories]
      summary: List all categories
      responses:
        "200":
          description: List of categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"
    post:
      operationId: createCategory
      tags: [categories]
      summary: Create a category
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCategoryRequest"
      responses:
        "201":
          description: Created category
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"

  /stats:
    get:
      operationId: getStats
      tags: [orders]
      summary: Get restaurant stats
      responses:
        "200":
          description: Restaurant statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Stats"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    TableStatus:
      type: string
      enum: [available, occupied, reserved, cleaning]

    Table:
      type: object
      properties:
        id:
          type: integer
        number:
          type: integer
        capacity:
          type: integer
        status:
          $ref: "#/components/schemas/TableStatus"
        guestName:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - number
        - capacity
        - status
        - createdAt
        - updatedAt

    CreateTableRequest:
      type: object
      properties:
        number:
          type: integer
        capacity:
          type: integer
      required:
        - number
        - capacity

    UpdateTableRequest:
      type: object
      properties:
        status:
          $ref: "#/components/schemas/TableStatus"
        guestName:
          type: string
          nullable: true
        capacity:
          type: integer

    OrderStatus:
      type: string
      enum: [pending, preparing, ready, delivered, cancelled]

    OrderItem:
      type: object
      properties:
        id:
          type: integer
        orderId:
          type: integer
        menuItemId:
          type: integer
        menuItemName:
          type: string
        menuItemPrice:
          type: number
        quantity:
          type: integer
        notes:
          type: string
          nullable: true
        subtotal:
          type: number
      required:
        - id
        - orderId
        - menuItemId
        - menuItemName
        - menuItemPrice
        - quantity
        - subtotal

    Order:
      type: object
      properties:
        id:
          type: integer
        tableId:
          type: integer
        tableNumber:
          type: integer
        status:
          $ref: "#/components/schemas/OrderStatus"
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        total:
          type: number
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - tableId
        - tableNumber
        - status
        - items
        - total
        - createdAt
        - updatedAt

    CreateOrderRequest:
      type: object
      properties:
        tableId:
          type: integer
        items:
          type: array
          items:
            $ref: "#/components/schemas/CreateOrderItemRequest"
        notes:
          type: string
          nullable: true
      required:
        - tableId
        - items

    CreateOrderItemRequest:
      type: object
      properties:
        menuItemId:
          type: integer
        quantity:
          type: integer
        notes:
          type: string
          nullable: true
      required:
        - menuItemId
        - quantity

    UpdateOrderRequest:
      type: object
      properties:
        status:
          $ref: "#/components/schemas/OrderStatus"
        notes:
          type: string
          nullable: true

    MenuItem:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
          nullable: true
        price:
          type: number
        categoryId:
          type: integer
        categoryName:
          type: string
        available:
          type: boolean
        preparationTime:
          type: integer
          description: Preparation time in minutes
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - price
        - categoryId
        - categoryName
        - available
        - createdAt

    CreateMenuItemRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
          nullable: true
        price:
          type: number
        categoryId:
          type: integer
        preparationTime:
          type: integer
          nullable: true
      required:
        - name
        - price
        - categoryId

    UpdateMenuItemRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
          nullable: true
        price:
          type: number
        categoryId:
          type: integer
        available:
          type: boolean
        preparationTime:
          type: integer
          nullable: true

    Category:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        icon:
          type: string
          nullable: true
        sortOrder:
          type: integer
      required:
        - id
        - name
        - sortOrder

    CreateCategoryRequest:
      type: object
      properties:
        name:
          type: string
        icon:
          type: string
          nullable: true
        sortOrder:
          type: integer
      required:
        - name
        - sortOrder

    Stats:
      type: object
      properties:
        totalOrders:
          type: integer
        totalRevenue:
          type: number
        averageTicket:
          type: number
        activeOrders:
          type: integer
        occupiedTables:
          type: integer
        totalTables:
          type: integer
        topItems:
          type: array
          items:
            $ref: "#/components/schemas/TopItem"
      required:
        - totalOrders
        - totalRevenue
        - averageTicket
        - activeOrders
        - occupiedTables
        - totalTables
        - topItems

    TopItem:
      type: object
      properties:
        menuItemId:
          type: integer
        name:
          type: string
        count:
          type: integer
        revenue:
          type: number
      required:
        - menuItemId
        - name
        - count
        - revenue