# lib/api-zod/src/generated/types/createOrderRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Comanda Eletrônica API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateOrderItemRequest } from "./createOrderItemRequest";

export interface CreateOrderRequest {
  tableId: number;
  items: CreateOrderItemRequest[];
  notes?: string | null;
}