# lib/api-zod/src/generated/types/menuItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Comanda Eletrônica API
 * OpenAPI spec version: 0.1.0
 */

export interface MenuItem {
  id: number;
  name: string;
  description?: string | null;
  price: number;
  categoryId: number;
  categoryName: string;
  available: boolean;
  /** Preparation time in minutes */
  preparationTime?: number | null;
  createdAt: Date;
}