# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Comanda Eletrônica API
 * OpenAPI spec version: 0.1.0
 */
import type { OrderItem } from "./orderItem";
import type { OrderStatus } from "./orderStatus";

export interface Order {
  id: number;
  tableId: number;
  tableNumber: number;
  status: OrderStatus;
  items: OrderItem[];
  total: number;
  notes?: string | null;
  createdAt: Date;
  updatedAt: Date;
}