# lib/api-zod/src/generated/types/stats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Comanda Eletrônica API
 * OpenAPI spec version: 0.1.0
 */
import type { TopItem } from "./topItem";

export interface Stats {
  totalOrders: number;
  totalRevenue: number;
  averageTicket: number;
  activeOrders: number;
  occupiedTables: number;
  totalTables: number;
  topItems: TopItem[];
}