# lib/api-zod/src/generated/types/tableStatus.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Comanda Eletrônica API
 * OpenAPI spec version: 0.1.0
 */

export type TableStatus = (typeof TableStatus)[keyof typeof TableStatus];

export const TableStatus = {
  available: "available",
  occupied: "occupied",
  reserved: "reserved",
  cleaning: "cleaning",
} as const;