# lib/api-zod/src/generated/types/topItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Comanda Eletrônica API
 * OpenAPI spec version: 0.1.0
 */

export interface TopItem {
  menuItemId: number;
  name: string;
  count: number;
  revenue: number;
}