# lib/api-zod/src/generated/types/updateMenuItemRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Comanda Eletrônica API
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateMenuItemRequest {
  name?: string;
  description?: string | null;
  price?: number;
  categoryId?: number;
  available?: boolean;
  preparationTime?: number | null;
}