# lib/api-zod/src/generated/types/updateTableRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Comanda Eletrônica API
 * OpenAPI spec version: 0.1.0
 */
import type { TableStatus } from "./tableStatus";

export interface UpdateTableRequest {
  status?: TableStatus;
  guestName?: string | null;
  capacity?: number;
}